backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
click==8.1.3
cmake==3.22.4
dlib==19.24.0
face_recognition==1.3.0
requests==2.27.1
opencv-python==4.5.5.64
Pillow==9.1.0
numpy==1.22.3
